requirements.txt
click==6.7
python-dotenv==0.7.1
requests==2.18.4
urllib3==1.22
